fix: removed useEffect for opacity update on scroll, opted for css only
